+++ map_real.h
@@ -1,6 +1,0 @@
-#pragma once
-#define RT_COUNT 277
-#define RT_COLS 32
-#define RT_ROWS 24
-extern const unsigned char RT_TILES[RT_COUNT][16];
-extern const unsigned short ROOM_NT[100][RT_ROWS][RT_COLS];